Switch to a full 64bit nonce

+++ core/src/commonJvmAndroidMain/kotlin/com/lightspark/sdk/core/crypto/SecureRandom.kt
@@ -2,6 +2,6 @@ package com.lightspark.sdk.core.crypto
 
 import java.security.SecureRandom
 
-internal actual fun nextInt(): Int {
-    return SecureRandom().nextInt()
+internal actual fun nextLong(): Long {
+    return SecureRandom().nextLong()
 }
